migrate TextField to NSTextView

+++ Code/macOS/Row.swift
@@ -46,7 +46,7 @@ class Row: NSTableRowView, Observer
         var drawRect = bounds
         
         let verticalGap = Float.verticalGap.cgFloat
-        drawRect.origin.y = verticalGap / 2
+        drawRect.origin.y = CGFloat(Int(verticalGap) / 2)
         drawRect.size.height -= verticalGap
         
         drawItemBackground(with: color, in: drawRect)
